restructured more of did
