+++ lets/src/id/identifier.rs
@@ -10,9 +10,6 @@ use async_trait::async_trait;
 // IOTA
 use crypto::{keys::x25519, signatures::ed25519};
 
-#[cfg(feature = "did")]
-use identity_iota::did::MethodScope;
-
 // Streams
 use spongos::{
     ddml::{
@@ -65,7 +62,10 @@ impl Identifier {
                 let doc = resolve_document(url_info).await?;
                 let method = doc
                     .document
-                    .resolve_method(url_info.exchange_fragment(), Some(MethodScope::key_agreement()))
+                    .resolve_method(
+                        url_info.exchange_fragment(),
+                        Some(identity_iota::did::MethodScope::key_agreement()),
+                    )
                     .expect("DID Method could not be resolved");
                 Ok(x25519::PublicKey::try_from_slice(&method.data().try_decode()?)?)
             }
